LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-target-margin-006-expected.txt

XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX TARGETTARGETTARGETTARGET XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX

PASS scroll-margin is taken into account when scrolling an inline element into view
